requirements.txt
PyMuPDF>=1.24
python-pptx>=0.6.23
